Refactored 2022 Day 1 to only test examples
