+++ 2022/day_01/Cargo.toml
@@ -3,8 +3,5 @@ name = "y22d01"
 version = "0.1.0"
 edition = "2021"
 
-[lib]
-doctest = false
-
 [dependencies]
 aoc_input = { path = "../../aoc_input" }
